fix: mode mainnet LH works
